Restructure PlugJS
