+++ package.json
@@ -2,12 +2,12 @@
   "name": "@croct/plug",
   "version": "0.0.0-dev",
   "description": "The hassle-free way to integrate Croct into any web application.",
+  "license": "MIT",
   "author": {
     "name": "Croct",
     "email": "plug-js+lib@croct.com",
     "url": "https://croct.com"
   },
-  "license": "UNLICENSED",
   "keywords": [
     "croct",
     "personalization",
@@ -15,8 +15,8 @@
     "javascript",
     "typescript"
   ],
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
+  "main": "index.js",
+  "types": "index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/croct-tech/plug-js.git"
@@ -26,40 +26,39 @@
   },
   "homepage": "https://github.com/croct-tech/plug-js",
   "scripts": {
-    "validate": "tsc --noEmit",
     "lint": "eslint 'src/**/*.ts' 'test/**/*.ts'",
     "test": "jest -c jest.config.js --coverage",
-    "rollup": "rollup -c rollup.config.js",
-    "rollup-min": "rollup -c rollup.min-config.js"
+    "validate": "tsc --noEmit",
+    "build": "tsc",
+    "bundle": "rollup -c rollup.config.js"
   },
   "dependencies": {
-    "@croct-tech/sdk": "0.1.4"
+    "@croct/sdk": "^0.2.0-alpha.4"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^11.1.0",
+    "@rollup/plugin-node-resolve": "^7.1.3",
+    "@rollup/plugin-typescript": "^4.1.1",
     "@types/jest": "^24.0.15",
     "@typescript-eslint/eslint-plugin": "^2.30.0",
     "@typescript-eslint/parser": "^2.30.0",
-    "dts-bundle-generator": "^3.3.1",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-jest": "^23.7.0",
     "jest": "^25.5.4",
-    "rollup": "^1.31.0",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-dts": "^1.4.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-typescript2": "^0.21.2",
+    "rollup": "^2.10.2",
+    "rollup-plugin-typescript2": "^0.27.1",
     "rollup-plugin-uglify": "^6.0.4",
     "temp-dir": "^2.0.0",
     "ts-jest": "^25.4.0",
-    "typescript": "^3.7.5"
+    "typescript": "^3.9.2"
   },
   "browserslist": [
     "last 1 version"
   ],
   "files": [
-    "build/index.js",
-    "build/index.d.ts"
+    "**/*.js",
+    "**/*.ts"
   ]
 }
