+++ src/index.ts
@@ -1,53 +1,4 @@
-import {
-    SdkFacade as Sdk,
-    Logger,
-    Tab,
-    UserFacade,
-    SessionFacade,
-    TrackerFacade as Tracker,
-    EvaluatorFacade as Evaluator,
-    EvaluationFacadeOptions as EvaluationOptions,
-    EvaluationErrorType,
-    EvaluationError,
-    ExpressionError,
-    Event,
-    EventType,
-    ExternalEvent,
-    ExternalEventPayload,
-    ExternalEventType,
-    EventListener,
-    EventInfo,
-    JsonValue,
-} from '@croct-tech/sdk';
-import {Plug, Configuration} from './plug';
-import {Plugin, PluginArguments, PluginFactory, PluginSdk} from './plugin';
+import {Configuration, Plug, GlobalPlug} from './plug';
 
-export {
-    Sdk,
-    Tab,
-    UserFacade,
-    SessionFacade,
-    Tracker,
-    Logger,
-    Evaluator,
-    EvaluationOptions,
-    Event,
-    EventType,
-    ExternalEventType,
-    ExternalEventPayload,
-    EventListener,
-    EventInfo,
-    ExternalEvent,
-    EvaluationErrorType,
-    EvaluationError,
-    ExpressionError,
-    JsonValue,
-    Plug,
-    Configuration,
-    PluginFactory,
-    PluginArguments,
-    Plugin,
-    PluginSdk,
-};
-
-export {default} from './singleton';
+export {Configuration, Plug};
+export default new GlobalPlug();
