+++ src/sdk/evaluation.ts
@@ -1,0 +1,2 @@
+export type {default as Evaluator, EvaluationOptions} from '@croct/sdk/facade/evaluatorFacade';
+export {EvaluationError, EvaluationErrorType, ExpressionError} from '@croct/sdk/evaluator';
