+++ src/sdk/tracking.ts
@@ -1,0 +1,2 @@
+export type {default as Tracker} from '@croct/sdk/facade/trackerFacade';
+export type {EventInfo, EventListener} from '@croct/sdk/tracker';
