--- src/singleton.ts
@@ -1,3 +1,0 @@
-import GlobalPlug from './globalPlug';
-
-export default new GlobalPlug();
